+++ secretflow/ml/nn/applications/sl_bst_torch.py
@@ -18,7 +18,7 @@ import torch
 from torch import nn as nn
 from torch.nn import functional as F
 
-from secretflow.ml.nn.utils import BaseModule
+from secretflow.ml.nn.core.torch import BaseModule
 
 
 class PositionalEmbedding(nn.Module):