+++ secretflow/ml/nn/fl/backend/tensorflow/strategy/fed_prox.py
@@ -15,8 +15,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-import collections
 import copy
 from typing import List, Tuple
 
@@ -65,26 +63,14 @@ class FedProx(BaseTFModel):
         logs = {}
 
         for _ in range(train_steps):
-            iter_data = next(self.train_set)
-            if len(iter_data) == 2:
-                x, y = iter_data
-                s_w = None
-            elif len(iter_data) == 3:
-                x, y, s_w = iter_data
-            if isinstance(x, collections.OrderedDict):
-                x = tf.stack(list(x.values()), axis=1)
-            num_sample += x.shape[0]
+            x, y, s_w = self.next_batch()
+            num_sample += self.get_sample_num(x)
 
             with tf.GradientTape() as tape:
                 # Step 1: forward pass
                 y_pred = self.model(x, training=True)
                 # Step 2: loss calculation, the loss function is configured in `compile()`.
-                loss = self.model.compiled_loss(
-                    y,
-                    y_pred,
-                    regularization_losses=self.model.losses,
-                    sample_weight=s_w,
-                )
+                loss = self.model.compute_loss(x, y, y_pred, s_w)
                 # assumption: the compiled loss is the estimated empirical loss on per single sample
                 if weights is not None:
                     # weights could be None in the very first step
@@ -98,7 +84,7 @@ class FedProx(BaseTFModel):
             gradients = tape.gradient(loss, trainable_vars)
             self.model.optimizer.apply_gradients(zip(gradients, trainable_vars))
             # Step4: update metrics
-            self.model.compiled_metrics.update_state(y, y_pred)
+            self.model.compute_metrics(x, y, y_pred, s_w)
         for m in self.model.metrics:
             logs[m.name] = m.result().numpy()
         self.wrapped_metrics.extend(self.wrap_local_metrics())