+++ tests/ml/nn/fl/model_def.py
@@ -14,6 +14,8 @@
 
 # Model zoo for unittest
 
+import numpy as np
+import tensorflow as tf
 import torch
 from torch import nn, optim
 from torch.nn import functional as F
@@ -181,3 +183,118 @@ class VAE(BaseModule):
     def configure_loss(self):
         """no use here, we use custom training step and loss calculation."""
         return None
+
+
+class CVAE(tf.keras.Model):
+    """Convolutional variational autoencoder."""
+
+    def __init__(self, latent_dim):
+        super(CVAE, self).__init__()
+        self.latent_dim = latent_dim
+        self.encoder = tf.keras.Sequential(
+            [
+                tf.keras.layers.InputLayer(input_shape=(28, 28, 1)),
+                tf.keras.layers.Conv2D(
+                    filters=32, kernel_size=3, strides=(2, 2), activation='relu'
+                ),
+                tf.keras.layers.Conv2D(
+                    filters=64, kernel_size=3, strides=(2, 2), activation='relu'
+                ),
+                tf.keras.layers.Flatten(),
+                # No activation
+                tf.keras.layers.Dense(latent_dim + latent_dim),
+            ]
+        )
+
+        self.decoder = tf.keras.Sequential(
+            [
+                tf.keras.layers.InputLayer(input_shape=(latent_dim,)),
+                tf.keras.layers.Dense(units=7 * 7 * 32, activation=tf.nn.relu),
+                tf.keras.layers.Reshape(target_shape=(7, 7, 32)),
+                tf.keras.layers.Conv2DTranspose(
+                    filters=64,
+                    kernel_size=3,
+                    strides=2,
+                    padding='same',
+                    activation='relu',
+                ),
+                tf.keras.layers.Conv2DTranspose(
+                    filters=32,
+                    kernel_size=3,
+                    strides=2,
+                    padding='same',
+                    activation='relu',
+                ),
+                # No activation
+                tf.keras.layers.Conv2DTranspose(
+                    filters=1, kernel_size=3, strides=1, padding='same'
+                ),
+            ]
+        )
+
+    def call(self, inputs, training=None, mask=None):
+        mean, logvar = self.encode(inputs)
+        return mean, logvar
+
+    @tf.function
+    def sample(self, eps=None):
+        if eps is None:
+            eps = tf.random.normal(shape=(100, self.latent_dim))
+        return self.decode(eps, apply_sigmoid=True)
+
+    def encode(self, x):
+        mean, logvar = tf.split(self.encoder(x), num_or_size_splits=2, axis=1)
+        return mean, logvar
+
+    def reparameterize(self, mean, logvar):
+        eps = tf.random.normal(shape=mean.shape)
+        return eps * tf.exp(logvar * 0.5) + mean
+
+    def decode(self, z, apply_sigmoid=False):
+        logits = self.decoder(z)
+        if apply_sigmoid:
+            probs = tf.sigmoid(logits)
+            return probs
+        return logits
+
+    def log_normal_pdf(self, sample, mean, logvar, raxis=1):
+        log2pi = tf.math.log(2.0 * np.pi)
+        return tf.reduce_sum(
+            -0.5 * ((sample - mean) ** 2.0 * tf.exp(-logvar) + logvar + log2pi),
+            axis=raxis,
+        )
+
+    def compute_loss(self, x=None, y=None, y_pred=None, sample_weight=None):
+        x = tf.expand_dims(x, axis=-1)
+        mean, logvar = self.encode(x)
+        z = self.reparameterize(mean, logvar)
+        x_logit = self.decode(z)
+        cross_ent = tf.nn.sigmoid_cross_entropy_with_logits(logits=x_logit, labels=x)
+        logpx_z = -tf.reduce_sum(cross_ent, axis=[1, 2, 3])
+        logpz = self.log_normal_pdf(z, 0.0, 0.0)
+        logqz_x = self.log_normal_pdf(z, mean, logvar)
+        return -tf.reduce_mean(logpx_z + logpz - logqz_x)
+
+    def compute_metrics(self, x, y_no_use, y_pred, sample_weight):
+        x = tf.expand_dims(x, axis=-1)
+        mean, logvar = y_pred
+        z = self.reparameterize(mean, logvar)
+        predictions = self.sample(z)
+        super().compute_metrics(None, x, predictions, sample_weight)
+
+    def get_config(self):
+        return {"latent_dim": self.latent_dim}
+
+
+def cvae_model():
+    def create_model():
+        model = CVAE(latent_dim=2)
+
+        model.compile(
+            optimizer=tf.keras.optimizers.Adam(1e-2),
+            loss=None,
+            metrics=[tf.keras.metrics.mean_squared_error],
+        )
+        return model
+
+    return create_model